refactor(transform): split orchestration entry along audited boundaries
