+++ packages/transform/src/__tests__/legacy-babel-reference/utils/processorLookup.ts
@@ -1,2 +1,1 @@
-export * from '../../../utils/processorLookup';
-
+export * from '../../../processors/processorLookup';
